+++ app/v2/definitions/filename_mapper.py
@@ -6,5 +6,3 @@ class FileNameMapperBase(ABC):
     @abstractmethod
     def get_output_type(self, filename: str, submission_type: str) -> str:
         pass
-
-
