+++ app/v2/definitions/location_key_lookup.py
@@ -1,7 +1,12 @@
 from abc import ABC, abstractmethod
+from typing import TypedDict
 
-from app.v2.definitions.config_schema import LocationKey
 from app.v2.definitions.location_name_repository import LookupKey
+
+
+class LocationKey(TypedDict):
+    location_type: str
+    location_name: str
 
 
 class LocationKeyLookupBase(ABC):
