+++ app/v2/location_key_lookup.py
@@ -1,7 +1,6 @@
 from typing import Final
 
-from app.v2.definitions.config_schema import LocationKey
-from app.v2.definitions.location_key_lookup import LocationKeyLookupBase
+from app.v2.definitions.location_key_lookup import LocationKeyLookupBase, LocationKey
 from app.v2.definitions.location_name_repository import LocationNameRepositoryBase, LookupKey
 
 # Location Types
